requirements.txt
langchain
langchain-community
langchain-openai
langgraph
streamlit
pydantic
python-dotenv
duckduckgo-search
tavily-python